# Generator: make `normalize_identifier` faster