@Kijewski 2024-01-11
I guess the benchmark for the different implementations does not need to be checked in, but I thought it might be good if I actually prove that my claims are not totally made up. :)

I would be interested if how the benchmark performs on different platforms, so if you have a non-AMD x64 CPU, or event a non-x64 CPU, then please post your results!