@Kijewski 2024-01-12
I optimized the implementation to compare arrays `[u8; _]` instead of slices `&[u8]`. The new `normalize_identifier_linear_by_len_replacement_only()` takes only 79.167µs on my machine.