@Kijewski 2024-01-12
I added a few comments, and made sure that all strings are ASCII.