@Kijewski 2024-01-12
Rust 1.60 is almost 2 years old. I guess we can raise the msrv, so I can use [`[u8].is_ascii()`](https://caniuse.rs/features/slice_escape_ascii), what do you say?